2016.4.24

+++ for 4/for 4.cpp
@@ -6,7 +6,7 @@
 
 int main()
 {
-	long int n, k;
+	int n, k;
 	scanf("%d", &n);
 	printf("%d=",n);
 	loop:for (k = 2; k < n; k++)
